==> notebooks/01_exploration.ipynb <==
import pandas as pd

df = pd.read_csv(r"E:\Projects CanadaIndia Data Scientist\credit-risk-project\data\credit_data.csv")
df.head()

df.info()
df.describe()

# What column represents default? SeriousDlqin2yrs
# Roughly what % of customers default?

(len(df[df["SeriousDlqin2yrs"]==1])/len(df))*100

# Only 6.68 % of data is defaulters
# Is the data imbalanced?
# Yes — the dataset is highly imbalanced.
# This is why: Accuracy is a bad metric. Metrics like ROC, KS, recall, and cutoff analysis matter more

df["RevolvingUtilizationOfUnsecuredLines"].value_counts()

# I can see one value 1.005733, ie, greater than 1 of RevolvingUtilizationOfUnsecuredLines
df["MonthlyIncome"].value_counts()

df.isna().sum()

df["DebtRatio"].value_counts()

'''
Target Variable (Not a Feature)
SeriousDlqin2yrs → Default (1 = default, 0 = no default)

A. Behavioral / Credit History Features
MOST predictive in real-world credit risk
-RevolvingUtilizationOfUnsecuredLines
How much of available credit is being used
Extremely strong risk signal
-NumberOfTime30-59DaysPastDueNotWorse
History of short-term delinquency
Strong early-warning indicator
-NumberOfTime60-89DaysPastDueNotWorse
More severe delinquency behavior
-NumberOfTimes90DaysLate
Very strong default predictor
-NumberOfOpenCreditLinesAndLoans
Credit activity and exposure
-NumberRealEstateLoansOrLines
Type and size of credit exposure
-NumberOfDependents (weak behavioral)
Financial burden proxy

B. Demographic / Profile Features
-age
Risk tends to decrease with age (generally)
-MonthlyIncome
Ability to repay, but often noisy / missing
-DebtRatio
Important but often distorted by missing income

C. Special / “Watchlist” Features (Very Important)
These features are useful but suspicious and require care.
-MonthlyIncome
Many missing values
Zero or extreme values common
-DebtRatio
Can explode when income is low or missing
-RevolvingUtilizationOfUnsecuredLines
Values > 1 possible (over-limit behavior)
'''
