==> notebooks/04_logistic_regression_baseline.ipynb <==
import pandas as pd

df_clean = pd.read_csv(r"E:\Projects Data Scientist\credit-risk-project\data\processed\credit_risk_clean.csv")

df_clean

# Define Features & Target
TARGET = "SeriousDlqin2yrs"

FEATURES = [
    "RevolvingUtilizationOfUnsecuredLines",
    "DebtRatio",
    "age",
    "MonthlyIncome",
    "NumberOfOpenCreditLinesAndLoans",
    "NumberRealEstateLoansOrLines",
    "TotalDelinquencyCount",
    "HighUtilizationFlag",
    "IncomeMissingFlag"
]

from sklearn.model_selection import train_test_split

X = df_clean[FEATURES]
y = df_clean[TARGET]

X_train, X_test, y_train, y_test = train_test_split(
    X, y,
    test_size=0.3,
    random_state=42,
    stratify=y
)

# Feature Scaling
from sklearn.preprocessing import StandardScaler

scaler = StandardScaler()

X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

# Train Logistic Regression (Class-Weighted)
from sklearn.linear_model import LogisticRegression

log_model = LogisticRegression(
    max_iter=1000,
    class_weight="balanced",
    solver="lbfgs"
)

log_model.fit(X_train_scaled, y_train)

# Generate Probability of Default Scores
y_train_pd = log_model.predict_proba(X_train_scaled)[:, 1]
y_test_pd = log_model.predict_proba(X_test_scaled)[:, 1]

# Interpret Coefficients
coef_df = pd.DataFrame({
    "Feature": FEATURES,
    "Coefficient": log_model.coef_[0]
}).sort_values(by="Coefficient", ascending=False)

coef_df

# Save the trained model
import joblib

joblib.dump(log_model, r"E:\Projects Data Scientist\credit-risk-project\models\logistic_pd_model.pkl")
joblib.dump(scaler, r"E:\Projects Data Scientist\credit-risk-project\models\standard_scaler.pkl")

test_results = X_test.copy()
test_results["actual_default"] = y_test.values
test_results["pd"] = y_test_pd

test_results.to_csv(r"E:\Projects Data Scientist\credit-risk-project\data\processed\test_scored.csv", index=False)

train_results = X_train.copy()
train_results["actual_default"] = y_train.values
train_results["pd"] = y_train_pd

train_results.to_csv(r"E:\Projects Data Scientist\credit-risk-project\data\processed\train_scored.csv", index=False)
